requirements.txt
python-telegram-bot[job-queue]==20.7 
selenium==4.15.2 
